requirements.txt
opencv-python==4.8.1.78
numpy==1.26.2
Pillow==10.1.0
pure-python-adb==0.3.0.dev0 